# POC: encoder.Encode() using IndentWriter