--- a/src/encoding/json/scanner_test.go
+++ b/src/encoding/json/scanner_test.go
@@ -9,6 +9,7 @@ import (
 	"math"
 	"math/rand"
 	"reflect"
+	"strings"
 	"testing"
 )
 
@@ -100,6 +101,27 @@ func TestCompactSeparators(t *testing.T) {
 			t.Errorf("Compact(%q) = %q, want %q", tt.in, s, tt.compact)
 		}
 	}
+}
+
+func benchmarkIndent(in string, b *testing.B) {
+	b.ReportAllocs()
+	b.SetBytes(int64(len(in)))
+	var buf bytes.Buffer
+	for n := 0; n < b.N; n++ {
+		buf.Reset()
+		_ = Indent(&buf, []byte(in), "", "\t")
+	}
+}
+
+func BenchmarkIndentExample1(b *testing.B) { benchmarkIndent(examples[0].compact, b) }
+func BenchmarkIndentExample2(b *testing.B) { benchmarkIndent(examples[1].compact, b) }
+func BenchmarkIndentExample3(b *testing.B) { benchmarkIndent(examples[2].compact, b) }
+func BenchmarkIndentExample4(b *testing.B) { benchmarkIndent(examples[3].compact, b) }
+func BenchmarkIndentExample5(b *testing.B) { benchmarkIndent(examples[4].compact, b) }
+func BenchmarkIndentExample6(b *testing.B) { benchmarkIndent(examples[5].compact, b) }
+func BenchmarkIndentExample7(b *testing.B) { benchmarkIndent(examples[6].compact, b) }
+func BenchmarkIndentExample7b(b *testing.B) {
+	benchmarkIndent(strings.Repeat(examples[6].compact, 100), b)
 }
 
 func TestIndent(t *testing.T) {