This PR is a POC for using the `IndentWriter` implementation in #1 to indent when using the `json.Encoder.SetIndent`, in addition to the new `json.Encoder.SetDirectWrite`, along with benchmarks to compare the before and after.

Before:

    BenchmarkMarshalAndIndentStream_Int-4            3000000               578 ns/op             176 B/op          3 allocs/op
    BenchmarkMarshalAndIndentStream_Optionals-4       500000              3265 ns/op             344 B/op          6 allocs/op
    BenchmarkMarshalAndIndentStream_StringTag-4       500000              2660 ns/op             336 B/op          6 allocs/op

After:

    BenchmarkMarshalAndIndentStream_Int-4            5000000               307 ns/op             224 B/op          3 allocs/op
    BenchmarkMarshalAndIndentStream_Optionals-4      2000000               798 ns/op             224 B/op          3 allocs/op
    BenchmarkMarshalAndIndentStream_StringTag-4      2000000               739 ns/op             232 B/op          4 allocs/op

So no improvement (and a slight memory hit) for really small values (a single int), but a clear improvement, both in terms of execution time, and memory usage/allocations, when using encoding and indenting structs.